Phase 1: extract features/diagram-export + harden harness for multi-line imports
- New markdown-viewer/src/features/diagram-export.js: downloadDiagramSvg/Png and
  their private helpers (getSvgElementForDiagram, triggerDownload). The
  fullscreen-overlay lookup is decoupled from main.js's cached element.
- Shared getSvgNaturalDimensions moves to core/utils.js (used by diagram fit,
  minimap, and export).
- Test harness now strips multi-line `import { ... } from '...'` statements
  (prettier wraps long imports), and detects relative deps via the `from`
  clause so the inliner follows the graph regardless of formatting.

Pure refactor. main.js ~2,786 -> ~2,706 lines. Build + lint clean; 297 tests pass.
