+++ markdown-viewer/src/main.js
@@ -14,7 +14,15 @@ import DOMPurify from 'dompurify';
 import 'highlight.js/styles/github-dark.css';
 
 // Internal modules (Phase 1 split — extracting cohesive units out of this entry).
-import { escapeHtml, normalizeMarkdownUrl } from './core/utils.js';
+import {
+  escapeHtml,
+  normalizeMarkdownUrl,
+  getSvgNaturalDimensions,
+} from './core/utils.js';
+import {
+  downloadDiagramSvg,
+  downloadDiagramPng,
+} from './features/diagram-export.js';
 
 // ===========================
 // Constants
@@ -1162,33 +1170,6 @@ function createDiagramContainer(svg, diagramId, mermaidSource) {
 // ===========================
 // Diagram Fit Helpers
 // ===========================
-function getSvgNaturalDimensions(svgElement) {
-    // SVG viewBox format: "min-x min-y width height"
-    // The 3rd and 4th values ARE the width and height (not coordinates)
-    const viewBox = svgElement.getAttribute('viewBox');
-    if (viewBox) {
-        const parts = viewBox.split(/[\s,]+/);
-        if (parts.length >= 4) {
-            const w = parseFloat(parts[2]);
-            const h = parseFloat(parts[3]);
-            if (w > 0 && h > 0) {
-                return { width: w, height: h };
-            }
-        }
-    }
-    // Fall back to width/height attributes (skip percentage values like "100%")
-    const wAttr = svgElement.getAttribute('width');
-    const hAttr = svgElement.getAttribute('height');
-    if (wAttr && hAttr && !String(wAttr).includes('%') && !String(hAttr).includes('%')) {
-        const w = parseFloat(wAttr);
-        const h = parseFloat(hAttr);
-        if (w > 0 && h > 0 && !isNaN(w) && !isNaN(h)) {
-            return { width: w, height: h };
-        }
-    }
-    return null;
-}
-
 function fitDiagramToContainer(wrapper, svgElement, panzoomInstance) {
     const dims = getSvgNaturalDimensions(svgElement);
     const containerWidth = wrapper.clientWidth;
@@ -2330,67 +2311,6 @@ function updateSplitRawPane(content) {
         .replace(/</g, '&lt;')
         .replace(/>/g, '&gt;');
     splitRawContent.innerHTML = `<code>${escaped}</code>`;
-}
-
-// ===========================
-// Feature: Diagram Export (SVG / PNG)
-// ===========================
-function getSvgElementForDiagram(diagramId) {
-    const wrapper = document.getElementById('wrapper-' + diagramId);
-    if (!wrapper) return null;
-    // Try original wrapper first, then fullscreen wrapper
-    return wrapper.querySelector('svg') ||
-        fullscreenOverlay.querySelector('.fullscreen-diagram-wrapper svg');
-}
-
-function downloadDiagramSvg(diagramId) {
-    const svgEl = getSvgElementForDiagram(diagramId);
-    if (!svgEl) return;
-
-    const serializer = new XMLSerializer();
-    const svgStr = serializer.serializeToString(svgEl);
-    const blob = new Blob([svgStr], { type: 'image/svg+xml;charset=utf-8' });
-    triggerDownload(blob, (diagramId || 'diagram') + '.svg');
-}
-
-function downloadDiagramPng(diagramId) {
-    const svgEl = getSvgElementForDiagram(diagramId);
-    if (!svgEl) return;
-
-    const serializer = new XMLSerializer();
-    const svgStr = serializer.serializeToString(svgEl);
-    const svgBlob = new Blob([svgStr], { type: 'image/svg+xml;charset=utf-8' });
-    const url = URL.createObjectURL(svgBlob);
-
-    const img = new Image();
-    img.onload = () => {
-        const canvas = document.createElement('canvas');
-        // Use natural SVG viewBox size for crisp export
-        const dims = getSvgNaturalDimensions(svgEl);
-        const scale = 2; // 2x for retina quality
-        canvas.width = (dims ? dims.width : img.naturalWidth || 800) * scale;
-        canvas.height = (dims ? dims.height : img.naturalHeight || 600) * scale;
-        const ctx = canvas.getContext('2d');
-        ctx.scale(scale, scale);
-        ctx.drawImage(img, 0, 0);
-        URL.revokeObjectURL(url);
-        canvas.toBlob((pngBlob) => {
-            triggerDownload(pngBlob, (diagramId || 'diagram') + '.png');
-        }, 'image/png');
-    };
-    img.onerror = () => URL.revokeObjectURL(url);
-    img.src = url;
-}
-
-function triggerDownload(blob, filename) {
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = filename;
-    document.body.appendChild(a);
-    a.click();
-    document.body.removeChild(a);
-    setTimeout(() => URL.revokeObjectURL(url), 1000);
 }
 
 // ===========================
